+++ Example/Utils/Config.swift
@@ -4,6 +4,7 @@ import MiniApp
 class Config: NSObject {
     enum Key: String {
         case applicationIdentifier = "RASApplicationIdentifier",
+        projectId = "RASProjectId",
         version = "CFBundleShortVersionString",
         subscriptionKey = "RASProjectSubscriptionKey",
         endpoint = "RMAAPIEndpoint",
@@ -13,8 +14,16 @@ class Config: NSObject {
     static let userDefaults = UserDefaults(suiteName: "com.rakuten.tech.mobile.miniapp.MiniAppDemo.settings")
 
     class func getCurrent() -> MiniAppSdkConfig {
+        guard let projectId = Config.userDefaults?.string(forKey: Config.Key.projectId.rawValue) else {
+            return MiniAppSdkConfig(baseUrl: Config.userDefaults?.string(forKey: Config.Key.endpoint.rawValue),
+                rasAppId: Config.userDefaults?.string(forKey: Config.Key.applicationIdentifier.rawValue),
+                subscriptionKey: Config.userDefaults?.string(forKey: Config.Key.subscriptionKey.rawValue),
+                hostAppVersion: Config.userDefaults?.string(forKey: Config.Key.version.rawValue),
+                isTestMode: Config.userDefaults?.bool(forKey: Config.Key.isPreviewMode.rawValue))
+        }
+
         return MiniAppSdkConfig(baseUrl: Config.userDefaults?.string(forKey: Config.Key.endpoint.rawValue),
-            rasAppId: Config.userDefaults?.string(forKey: Config.Key.applicationIdentifier.rawValue),
+            rasProjectId: projectId,
             subscriptionKey: Config.userDefaults?.string(forKey: Config.Key.subscriptionKey.rawValue),
             hostAppVersion: Config.userDefaults?.string(forKey: Config.Key.version.rawValue),
             isPreviewMode: Config.userDefaults?.bool(forKey: Config.Key.isPreviewMode.rawValue))