+++ Tests/Unit/Helpers.swift
@@ -18,7 +18,7 @@ class MockAPIClient: MiniAppClient {
                     rasAppId: bundle.mockAppId,
                     subscriptionKey: bundle.mockSubscriptionKey,
                     hostAppVersion: bundle.mockHostAppUserAgentInfo,
-                    isPreviewMode: bundle.mockTestMode
+                    isTestMode: bundle.mockTestMode
                 )
         )
     }
@@ -168,6 +168,7 @@ class MockBundle: EnvironmentProtocol {
 
     var mockValueNotFound: String?
     var mockAppId: String?
+    var mockProjectId: String?
     var mockSubscriptionKey: String?
     var mockAppVersion: String?
     var mockEndpoint: String? = "https://www.example.com/"
@@ -187,6 +188,8 @@ class MockBundle: EnvironmentProtocol {
         switch key {
         case "RASApplicationIdentifier":
             return mockAppId
+        case "RASProjectId":
+            return mockProjectId
         case "RASProjectSubscriptionKey":
             return mockSubscriptionKey
         case "CFBundleShortVersionString":